+++ TF_Settings_Web/src/Components/Header/index.ts
@@ -1,6 +1,5 @@
-import Header from './Header/Header';
 import StatusIndicator from './StatusIndicator/StatusIndicator';
 import TabBar from './TabBar/TabBar';
 import TabSelector from './TabSelector/TabSelector';
 
-export { Header, StatusIndicator, TabSelector, TabBar };
+export { StatusIndicator, TabSelector, TabBar };
